requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
djangorestframework==3.14.0
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
